put wrapper in separate file, add comments, clean
